tasks/vision/classification/MNasNet/requirements.txt
tensorflow==2.13.0
numpy==1.24.0
Pillow==10.0.0
tqdm==4.64.1
tensorflow_datasets==4.9.2